feat: Rename "errors" directory -> "Error"
