+++ InstanceRenderer.js
@@ -1,5 +1,5 @@
 import {WebGLRenderer} from "./WebGLRenderer.js";
-import {NoWebGL2Error} from "./errors/index.js";
+import {NoWebGL2Error} from "./Error/index.js";
 import {ShaderLoader} from "./Loader/index.js";
 
 /**
