+++ gui/Layer.js
@@ -1,6 +1,6 @@
 import {GUIComposite} from "./index.js";
 import {Component} from "./Component/index.js";
-import {NotImplementedError} from "../errors/index.js";
+import {NotImplementedError} from "../Error/index.js";
 
 /**
  * @todo Replace by Component?
